+++ cartsy/ViewController Extensions.swift
@@ -10,9 +10,15 @@ import Foundation
 import UIKit
 
 extension UIViewController {
-    /// when
+    /// + Button Pressed
+    ///
+    /// :param: object String representation of what you're saving, only used in alert dialogue
+    /// :param: inTable UITableView of current View to reload data.
+    /// :param: savedAs Save function for this type so that we can handle Lists and Items differents
+    ///
+    /// :returns: Void. Calls savedAs function and reloads view.
     func addNewPressed(#object: String, inTable: UITableView, savedAs: (String) -> Void) { /// TODO: try to make an optional closure at the end, should we need it
-        // make alert with a save/cancel dialog
+        // make alert with a save/cancel dialogue
         var alert = UIAlertController(title: "Add", message: "Add new \(object)", preferredStyle: .Alert)
         let saveAction = UIAlertAction(title: "Save", style: .Default) { (action: UIAlertAction!) -> Void in
             let textField = alert.textFields![0] as UITextField